requirements.txt
psutil==5.6.3
py-cpuinfo==5.0.0